+++ R/zonal_stats.R
@@ -1,16 +1,22 @@
-#' Get summary of zonal statistics
+#' Get zonal statistics
+#'
+#' Get zonal statistics. For *each* of the `n_days` prior to the sample event date,
+#' returns summarised values of covariate data for `radius` metres around each site
+#' location, using `spatial_stats` to determine how to spatially summarise the data.
 #'
 #' @param se Sample events from \code{mermaidr}
 #' @param covariate Covariate to get statistics for. Both covariate title or ID are permitted.
 #' @param n_days Number of days to get statistics for. Includes the sample date
 #'  itself, and days prior to it -- e.g., 365 days would include the sample date
 #'  and the 364 days prior. Defaults to 365.
-#' @param buffer Buffer around site location, in metres. Defaults to 1000.
-#' @param stats Summary statistics. One of: min, max, or mean.
+#' @param radius Radius around site location, in metres. Defaults to 1000.
+#' @param spatial_stats Spatial statistics -- used to summarise all data around
+#' the site location, according to the \code{radius} set.
 #'
 #' @export
 get_zonal_statistics <- function(se, covariate, n_days = 365,
-                                 buffer = 1000, stats = c("min", "max", "mean")) {
+                                 radius = 1000,
+                                 spatial_stats = c("min", "max", "mean")) {
   covariate_id <- get_covariate_id(covariate)
 
   if (covariate_id == covariate) {
@@ -19,85 +25,23 @@ get_zonal_statistics <- function(se, covariate, n_days = 365,
     covariate_name <- covariate
   }
 
-  # Add an ID for iterating over (with site/date/lat/long distinct)
-  se <- se %>%
-    add_id_for_iteration()
+  if (!"...id" %in% names(se)) { # Don't need to do this if get_summary_zonal_statistics()
+    # called get_zonal_
+
+    # Add an ID for iterating over (with site/date/lat/long distinct)
+    se <- se %>%
+      add_id_for_iteration()
+  }
 
   se_list <- se %>%
     split(.$...id)
 
   # Get zonal stats for all SEs
-  zonal_stats <- get_zonal_stats(se_list, covariate_id, n_days, buffer, stats)
+  zonal_stats <- get_zonal_stats(se_list, covariate_id, n_days, radius, spatial_stats)
 
-  # Set up to group by non-stat columns
-  id_cols <- zonal_stats %>%
-    dplyr::select(-dplyr::any_of(stats)) %>%
-    names()
-
-  # Apply summary function to each of the cols
-  zonal_stats_summary <- stats %>%
-    purrr::map(\(x) {
-      zonal_stats %>%
-        dplyr::group_by(dplyr::across(dplyr::all_of(id_cols))) %>%
-        dplyr::summarise(
-          dplyr::across(
-            dplyr::any_of(x),
-            ~ {
-              if (x == "mean") {
-                if (all(is.na(.x))) {
-                  NA
-                } else {
-                  mean(.x, trim = 0, na.rm = TRUE)
-                }
-              } else {
-                do.call(x, as.list(.x, na.rm = TRUE))
-              }
-            }
-          ),
-          .groups = "drop"
-        )
-    }) %>%
-    dplyr::bind_cols()
-
-  # Reshape zonal stats into the following format:
-  # covariate, start_date, end_date, band, statistic, value
-  # covariate will just be covariate name
-  # band is as is, with "_band" removed
-  # Put into a df-column called covariates
-
-  if (!any(stats %in% names(zonal_stats_summary))) {
-    zonal_stats_summary <- zonal_stats_summary %>%
-      dplyr::bind_cols(dplyr::tibble(
-        statistic = NA_character_,
-        value = NA_real_,
-        band = NA_character_
-      ))
-  } else {
-    zonal_stats_summary <- zonal_stats_summary %>%
-      tidyr::pivot_longer(
-        cols = dplyr::any_of(stats),
-        names_to = "statistic",
-        values_to = "value"
-      ) %>%
-      dplyr::mutate(
-        band = stringr::str_remove(band, "band_"),
-        band = as.numeric(band)
-      )
-  }
-
-  zonal_stats_df <- zonal_stats_summary %>%
-    dplyr::right_join(se, by = "...id") %>%
-    dplyr::mutate(
-      covariate = covariate_name,
-      start_date = start_date,
-      end_date = end_date
-    ) %>%
-    dplyr::select(...id, covariate, start_date, end_date, n_dates, band, statistic, value) %>%
-    tidyr::nest(covariates = -...id)
-
-  # Re-attach to existing df, even if it was not distinct
+  # Attach to sample events and remove ID
   se %>%
-    dplyr::left_join(zonal_stats_df, by = "...id") %>%
+    dplyr::left_join(zonal_stats, by = "...id") %>%
     dplyr::select(-...id)
 }
 
@@ -177,15 +121,15 @@ get_items_for_zonal_stats <- function(df, covariate_id, n_days = 365) {
 }
 
 
-get_zonal_stats <- function(se_list, covariate_id, n_days, buffer, stats) {
+get_zonal_stats <- function(se_list, covariate_id, n_days, radius, spatial_stats) {
   se_list %>%
     purrr::map(
       \(se)
       safely_get_zonal_stats_single(se,
         covariate_id,
         n_days,
-        buffer = buffer,
-        stats = stats
+        radius = radius,
+        spatial_stats = spatial_stats
       ),
       .progress = TRUE
     ) %>%
@@ -193,12 +137,12 @@ get_zonal_stats <- function(se_list, covariate_id, n_days, buffer, stats) {
     purrr::compact() %>% # Remove those without any items/results
     purrr::list_rbind(names_to = "...id") %>%
     dplyr::mutate(n_dates = dplyr::n()) %>%
-    dplyr::select(-date)
+    tidyr::nest(covariates = -...id, .by = "...id") # Nest covariates
 }
 
-get_zonal_stats_single <- function(se, covariate_id, n_days = 30, buffer = 1000,
+get_zonal_stats_single <- function(se, covariate_id, n_days = 30, radius = 1000,
                                    bands = list(1), approx_stats = FALSE,
-                                   stats = c(
+                                   spatial_stats = c(
                                      "min", "max", "mean", "count", "sum", "std",
                                      "median", "majority", "minority", "unique",
                                      "range", "nodata", "area", "freq_hist"
@@ -229,10 +173,10 @@ get_zonal_stats_single <- function(se, covariate_id, n_days = 30, buffer = 1000,
       aoi = list(
         type = "Point",
         coordinates = c(se[["longitude"]], se[["latitude"]]),
-        buffer_size = buffer
+        buffer_size = radius
       ),
       image = NULL,
-      stats = as.list(stats)
+      stats = as.list(spatial_stats)
     )) %>%
     httr2::req_error(is_error = \(res) FALSE)
 
@@ -275,13 +219,25 @@ get_zonal_stats_single <- function(se, covariate_id, n_days = 30, buffer = 1000,
             x <- purrr::map(x, \(x) if (is.null(x)) NA else x)
             dplyr::as_tibble(x)
           }, .id = "band") %>%
-          dplyr::mutate(date = date)
+          # This should happen further out
+          dplyr::mutate(
+            date = date,
+            band = stringr::str_remove(band, "band_"),
+            band = as.numeric(band)
+          )
       }
     ) %>%
     purrr::list_rbind() %>%
-    dplyr::bind_cols(stac_items %>% dplyr::distinct(start_date, end_date))
+    dplyr::bind_cols(stac_items %>% dplyr::distinct(start_date, end_date)) %>%
+    # This should happen further out
+    tidyr::pivot_longer(
+      cols =
+        dplyr::any_of(spatial_stats),
+      names_to = "spatial_stat",
+      values_to = "value"
+    )
+  # TODO -> reorder cols too, ALSO further out
 }
-
 
 safely_get_zonal_stats_single <- purrr::safely(get_zonal_stats_single, otherwise = NULL)